requirements.txt
# Core
Django==4.2.7

# Database
psycopg2-binary>=2.9.9,<3.0.0  # PostgreSQL adapter
dj-database-url>=2.1.0,<3.0.0  # Database URL parsing
python-dotenv>=1.0.0,<2.0.0    # Environment variable management

# ASGI Server
gunicorn==21.2.0                # Production WSGI server
whitenoise==6.6.0              # Static file serving

# REST Framework
djangorestframework>=3.15.0,<4.0.0
drf-nested-routers>=0.95.0,<1.0.0
djangorestframework-simplejwt>=5.3.1,<6.0.0

# API Documentation
drf-yasg==1.21.7

# Security
python-dotenv==1.0.0
PyJWT==2.8.0

# CORS
django-cors-headers==4.3.1

# Static files
whitenoise==6.6.0

# ASGI Server
gunicorn==21.2.0
uvicorn==0.23.2

# Image Processing
# Using Pillow 10.2.0+ for Python 3.13 compatibility
Pillow>=10.2.0,<11.0.0

# Utilities
pytz==2023.3.post1
sqlparse==0.4.4

# Development (not needed in production)
pip==23.2.1
setuptools==68.0.0
wheel==0.40.0
